client/python/examples/requirements.txt
click==8.0.3
pensando_dss
python-dateutil==2.8.2
PyYAML==6.0
requests==2.26.0
urllib3==1.26.6
